+++ .mvn/wrapper/maven-wrapper.properties
@@ -1,0 +1,1 @@
+distributionUrl=https://repo1.maven.org/maven2/org/apache/maven/apache-maven/3.1.0/apache-maven-3.1.0-bin.zip
