+++ noah-core/noah-framework/src/main/java/com/alibaba/rigel/framework/FrameworkConstants.java
@@ -2,8 +2,6 @@ package com.alibaba.rigel.framework;
 
 public interface FrameworkConstants {
 	
-	static final String APP = "lotus";
-	
 	static final String CHARSET = "UTF-8";
 	
 	public static final String REQUEST_ID = "request_id";
