+++ noah-core/noah-framework/src/main/java/com/alibaba/rigel/framework/template/Renderable.java
@@ -1,0 +1,5 @@
+package com.alibaba.rigel.framework.template;
+
+public interface Renderable {
+	String render();
+}
