+++ noah-web/src/main/resources/applicationContext.xml
@@ -1,26 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:context="http://www.springframework.org/schema/context" xmlns:util="http://www.springframework.org/schema/util"
-	xsi:schemaLocation="
+    xmlns:context="http://www.springframework.org/schema/context" xmlns:util="http://www.springframework.org/schema/util"
+    xsi:schemaLocation="
         http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util.xsd
         http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd
         http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-	<context:property-placeholder location="classpath:conf/dev.properties" />
+    <context:property-placeholder location="classpath:conf/dev.properties" />
 
-	<bean name="messageSource" class="org.springframework.context.support.ReloadableResourceBundleMessageSource">
-		<property name="defaultEncoding" value="utf-8" />
-		<property name="cacheSeconds" value="30" />
-		<property name="basenames">
-			<list>
-				<value>/WEB-INF/conf/system</value>
-				<value>/WEB-INF/conf/url</value>
-				<value>/WEB-INF/conf/index</value>
-			</list>
-		</property>
-	</bean>
+    <bean name="messageSource" class="org.springframework.context.support.ReloadableResourceBundleMessageSource">
+        <property name="defaultEncoding" value="utf-8" />
+        <property name="cacheSeconds" value="30" />
+        <property name="basenames">
+            <list>
+                <value>/WEB-INF/conf/system</value>
+                <value>/WEB-INF/conf/url</value>
+                <value>/WEB-INF/conf/index</value>
+            </list>
+        </property>
+    </bean>
 
-	<import resource="applicationContext-biz.xml" />
-	<import resource="applicationContext-common.xml" />
-	<import resource="applicationContext-db.xml" />
+    <import resource="applicationContext-biz.xml" />
+    <import resource="applicationContext-common.xml" />
+    <import resource="applicationContext-db.xml" />
 </beans>
