+++ noah-web/src/main/webapp/WEB-INF/templates/module/index.vm
@@ -22,7 +22,7 @@ button {
 <br />
 $requestContext.getMessage("var.copyright")
 <script>
-	var dataList = $JsonUtils.stringify($user); 
+	var dataList = $JsonUtils.stringify($user); 	
 </script>
 <input name="request_id" value="$request_id" />
 <input name="session_id" value="$session.id" />
